@@ -5,7 +5,7 @@ author: flyzero
 homepage: https://github.com/flymzero/webdav_client
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.10.0 <3.0.0"
 
 dependencies:
   crypto: ^2.1.5
